requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.2
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.1
cssselect==1.2.0
daphne==4.0.0
Django==3.2.19
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.14.0
django-filter==23.2
django-grip==3.3.0
django-multimedia==0.2.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ffmpeg-python==0.2.0
future==0.18.3
gripcontrol==4.1.0
gunicorn==21.1.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.13.0
incremental==22.10.0
kombu==5.2.4
lxml==4.9.2
Markdown==3.4.3
markdown2==2.4.8
MarkupSafe==2.1.3
msgpack==1.0.5
packaging==23.1
paramiko==3.2.0
Pillow==9.5.0
prompt-toolkit==3.0.38
pubcontrol==3.4.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyquery==2.0.0
python-dotenv==0.21.1
pytz==2023.3
readtime==2.0.0
redis==4.5.1
requests==2.31.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
storage==0.0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing-extensions==4.5.0
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.15.0
zope.interface==6.0
